# src/main/resources/openapi/UserManagement.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User Management API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9150/user-management/api"
    }
  ],
  "paths": {
    "/user": {
      "post": {
        "summary": "Create user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/User.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/User.json"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "description": "Create user"
      }
    },
    "/service": {
      "post": {
        "summary": "Create service",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/Service.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/Service.json"
                }
              }
            }
          }
        },
        "tags": [
          "Service"
        ],
        "description": "Create service"
      },
      "get": {
        "summary": "Find services by exact name (case-sensitive)",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/Service.json"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Service"
        ],
        "description": "Find services by exact name (case-sensitive)"
      }
    },
    "/subscription": {
      "post": {
        "summary": "Create subscription",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/Subscription.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/Subscription.json"
                }
              }
            }
          }
        },
        "tags": [
          "Subscription"
        ],
        "description": "Create subscription"
      }
    },
    "/subscription-user": {
      "post": {
        "summary": "Create subscription user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/SubscriptionUser.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/SubscriptionUser.json"
                }
              }
            }
          }
        },
        "tags": [
          "SubscriptionUser"
        ],
        "description": "Create subscription user"
      }
    },
    "/subscription-user/subscription/{subscriptionId}/user/{userId}/permissions": {
      "get": {
        "summary": "Get effective permissions for a subscription user",
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/PermissionList.json"
                }
              }
            }
          }
        },
        "tags": [
          "SubscriptionUser"
        ],
        "description": "Get effective permissions for a subscription user"
      }
    },
    "/subscription-user/subscription/{subscriptionId}/user/{userId}/special-permissions": {
      "post": {
        "summary": "Upsert special permissions for a subscription user",
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "./schemas/SpecialPermission.json"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/SubscriptionUser.json"
                }
              }
            }
          }
        },
        "tags": [
          "SubscriptionUser"
        ],
        "description": "Upsert special permissions for a subscription user"
      }
    },
    "/user/{userId}/permissions": {
      "get": {
        "summary": "Get all permissions for a user across subscriptions",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schemas/UserPermission.json"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "description": "Get all permissions for a user across subscriptions"
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "$ref": "./schemas/User.json"
      },
      "Service": {
        "$ref": "./schemas/Service.json"
      },
      "Subscription": {
        "$ref": "./schemas/Subscription.json"
      },
      "SubscriptionUser": {
        "$ref": "./schemas/SubscriptionUser.json"
      },
      "SpecialPermission": {
        "$ref": "./schemas/SpecialPermission.json"
      },
      "PermissionList": {
        "$ref": "./schemas/PermissionList.json"
      },
      "Role": {
        "$ref": "./schemas/Role.json"
      },
      "Permission": {
        "$ref": "./schemas/Permission.json"
      },
      "RolePermissionMapping": {
        "$ref": "./schemas/RolePermissionMapping.json"
      },
      "UserPermission": {
        "$ref": "./schemas/UserPermission.json"
      }
    }
  }
}